Rework state components

See merge request swipe/swipe!156

+++ backend/authorization/views.py
@@ -27,6 +27,8 @@ class Logout(View):
         return super(Logout, self).dispatch(request, *args, **kwargs)
 
     def post(self, request):
+        if 'token' not in request.POST or request.POST['token'] is None:
+            return JSONResponse({})
         Token.objects.filter(key=request.POST['token']).delete()
         return JSONResponse({})
